Split new classes into separate files
