+++ NuKeeper/GitHub/IGitHubAccount.cs
@@ -1,0 +1,9 @@
+namespace NuKeeper.GitHub
+{
+    public interface IGitHubAccount
+    {
+        string Login { get; }
+        string Name { get; }
+        string Email { get; }
+    }
+}
