+++ NuKeeper/GitHub/IGitHubRepositoryPermissions.cs
@@ -1,0 +1,8 @@
+namespace NuKeeper.GitHub
+{
+    public interface IGitHubRepositoryPermissions
+    {
+        bool Push { get; }
+        bool Pull { get; }
+    }
+}
